+++ SourceCode/Data/ModuleStandard.cs
@@ -15,6 +15,7 @@ public partial class ModuleStandard
     public string Electricity { get; set; }
     public string PreferredTheme { get; set; }
     public string AcceptedNorm { get; set; }
+    public string MainTheme { get; set; } = "EUROPE";
 
     public virtual Scale Scale { get; set; }
 }