+++ SourceCode/Database/dbo/Procedures/AddGeneratedModuleWaybills.sql
@@ -26,10 +26,11 @@ BEGIN
 	WHERE
 		ME.StationCustomerId = @StationCustomerId
 		AND ME.StationCustomerId <> OTHER.StationCustomerId
-		AND ME.IsSupply <> OTHER.IsSupply
-		--AND ME.QuantityUnitId = OTHER.QuantityUnitId
+		AND ME.IsSupply <> OTHER.IsSupply AND ME.IsSupply = 0
+		--AND ME.QuantityUnitId = OTHER.QuantityUnitId -- Is not used, because it causes too few matches.
 		AND ME.StationId <> OTHER.StationId
-		AND ME.ScaleId = OTHER.ScaleId
+		AND ME.MainTheme = OTHER.MainTheme
+		AND (ME.ScaleId = OTHER.ScaleId OR ME.ScaleId = 0 OR OTHER.ScaleId = 0)
 		AND (ME.CountryId = OTHER.CountryId OR ME.IsInternational <> 0 AND OTHER.IsInternational <>0) 
 		AND (ME.FromYear IS NULL OR OTHER.UptoYear IS NULL OR ME.FromYear <= OTHER.UptoYear )
 		AND (ME.UptoYear IS NULL OR OTHER.FromYear IS NULL OR ME.UptoYear >= OTHER.FromYear )