+++ SourceCode/Database/dbo/Tables/ModuleStandard.sql
@@ -10,6 +10,7 @@
     [Electricity]    NVARCHAR (20)  NULL,
     [PreferredTheme] NVARCHAR (50)  NULL,
     [AcceptedNorm]   NVARCHAR (255) NULL,
+    [MainTheme] NVARCHAR(10) NULL, 
     CONSTRAINT [PK_ModuleStandard] PRIMARY KEY CLUSTERED ([Id] ASC),
     CONSTRAINT [FK_ModuleStandard_Scale] FOREIGN KEY ([ScaleId]) REFERENCES [dbo].[Scale] ([Id])
 );